requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gTTS==2.3.2
h5py==3.9.0
idna==3.4
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
libretranslatepy==2.1.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.10.3
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
shiboken6==6.5.2
six==1.16.0
sounddevice==0.4.6
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
translate==3.6.1
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
